Top and bottom chemicals for partition settler
This can help the users specify which chemicals will definitely stay in
one of the phases and the rest will go into the other. Useful when
composing systems and you won't know everything about the composition of
the feed before simulating the complete system.

Maybe a process specification can be an alternative to this.